pyproject.toml
[project]
name = "Boom-Tetris"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.5",
    "pygame>=2.6.1",
    "pyyaml>=6.0.2",
    "ruff>=0.11.6",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
